python test: move single-use fixtures to subdirectories
